[#58] Controller 예외처리 통합
