+++ src/main/java/com/snack/news/exception/base/BadRequestException.java
@@ -4,7 +4,7 @@ import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ResponseStatus;
 
 @ResponseStatus(value = HttpStatus.BAD_REQUEST)
-public class BadRequestException extends IllegalArgumentException {
+public abstract class BadRequestException extends RuntimeException {
 	public final static String ERROR_CODE = "BAD_REQUEST";
 
 	public String getErrorCode() {
