+++ src/test/resources/import.sql
@@ -1,4 +1,4 @@
--- Category 테스트 데이터
+    -- Category 테스트 데이터
 insert into category (id, title) values (1, 'IT');
 insert into category (id, title) values (2, '커머스');
 
@@ -17,7 +17,7 @@ insert into news(id, category_id, title, content, create_at, modified_at) values
 insert into news(id, category_id, title, content, create_at, modified_at) values (2, 2, '두 번째 뉴스 타이틀', '두 번째 뉴스 내용입니다.', '2019-07-21 00:00:00', '2019-10-21 00:00:00');
 insert into news(id, category_id, title, content, create_at, modified_at) values (3, 1, '세 번째 뉴스 타이틀', '세 번째 뉴스 내용입니다.', '2019-08-21 00:00:00', '2019-10-21 00:00:00');
 insert into news(id, category_id, title, content, create_at, modified_at) values (4, 2, '네 번째 뉴스 타이틀', '네 번째 뉴스 내용입니다.', '2019-09-21 00:00:00', '2019-10-21 00:00:00');
-insert into news(id, category_id, title, content, create_at, modified_at) values (5, 2, '다섯 번째 뉴스 타이틀', '다섯 번째 뉴스 내용입니다.', '2019-07-21 00:00:00', '2019-10-21 00:00:00');
+insert into news(id, category_id, title, content, create_at, modified_at) values (5, 2, '다섯 번째 뉴스 타이틀', '다섯 번째 뉴스 내용입니다.', '2019-07-23 00:00:00', '2019-10-21 00:00:00');
 
 -- News-Topic 연관관계 테스트 데이터
 insert into news_topic(news_id, topic_id) values (1, 1);
